.......... [DEV-3899] changes after review

+++ tests/libs/zbxstr/zbx_strcmp_null.c
@@ -25,7 +25,7 @@ void	zbx_mock_test_entry(void **state)
 	const char	*s1 = zbx_mock_get_parameter_string("in.string1");
 	const char	*s2 = zbx_mock_get_parameter_string("in.string2");
 	int			exp_result = zbx_mock_str_to_return_code(zbx_mock_get_parameter_string("out.val"));
-	size_t		act_value = zbx_strcmp_null(s1, s2);
+	int		act_value = zbx_strcmp_null(s1, s2);
 
 	zbx_mock_assert_int_eq("return value",  exp_result, act_value);
 }
